Add every serotype reconstructed root and midpoint
Using the cis-serotype reconstructed root as the root of the tree causes
the outgroups to be be assigned to the same clade as the root.
(e.g. A "DENV1" root will cause the outgroups (DENV2, DENV3, DENV4) to be
assigned to the DENV1 clade.)

Instead include all the serotype reconstructed roots and have augur refine
midpoint the tree such that the root of the tree is "unassigned" and we retain
our outgroups being "unassigned".